Fix changed location of reset_frame_range
Location in commands caused cyclic import
